requirements.txt
httpx
python-dotenv
pandas